Consider a cell eol if there are no more cells to the right.
Some applications seem to emit null characters instead spaces. With the previous
logic of considering a cell eol if all cells to its right are null makes it so
that typing spaces in applications such as Vim or MySQL client have no effect on
the location of the cursor, making entering input more difficult.

One observed downside with this change is that executing vterm-end-of-line in
vterm-copy-mode moves the cursor all the way to the right edge of the line
instead of to the last non-null character.